move all utility functions to separate file
